+++ src/agedi/cli/main.py
@@ -4,6 +4,7 @@ from agedi.cli.train_hydra import train_hydra
 from agedi.cli.sample import sample
 from agedi.cli.inspect import inspect
 from agedi.cli.predict import predict
+from agedi.cli.relax import relax
 
 @click.group()
 @click.version_option()
@@ -16,4 +17,5 @@ cli.add_command(train_hydra)
 cli.add_command(sample)
 cli.add_command(inspect)
 cli.add_command(predict)
+cli.add_command(relax)